LayoutTests/editing/selection/ios/selection-highlight-in-fixed-position-container.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true SelectionHonorsOverflowScrolling=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta charset="utf-8">
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
body, html {
    font-size: 16px;
    font-family: system-ui;
    height: 3000px;
    margin: 0;
}

.container {
    width: 100%;
    height: 100px;
    top: 0;
    background: tomato;
    color: white;
    position: fixed;
    text-align: center;
    font-size: 32px;
    line-height: 32px;
}
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("Verifies that the selection highlight inside a fixed-position container remains visible while scrolling");

    await UIHelper.longPressElement(document.getElementById("target"));
    await UIHelper.waitForSelectionToAppear();
    testPassed("Selected text");

    await UIHelper.sendEventStream(new UIHelper.EventStreamBuilder()
        .begin(150, 400)
        .move(150, 80, 0.25)
        .end()
        .takeResult());

    selectionRects = await UIHelper.getUISelectionViewRects();
    shouldBe("selectionRects.length", "1");
    await UIHelper.waitForZoomingOrScrollingToEnd();
    finishJSTest();
});
</script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
    <div class="container">
        <p><span id="target">Select</span> me</p>
    </div>
</body>
</html>
